implemented validations

+++ src/Application/Controller/BlogController.php
@@ -6,26 +6,30 @@ namespace App\Application\Controller;
 use App\Application\Commands\ArticlePublishCommand;
 use App\Application\Commands\BlogCreateCommand;
 use App\Application\Commands\BlogEditCommand;
-use App\Application\Commands\Handler\ArticlePublisherHandle;
+use App\Application\Commands\Handler\ArticlePublisherHandler;
 use App\Application\Commands\Handler\BlogCreateHandler;
 use App\Application\Commands\Handler\BlogEditHandler;
 use App\Domain\Model\Type\BlogId;
 use App\Domain\Repository\BlogRepositoryInterface;
 use App\Domain\Repository\Exception\BlogNotFoundException;
 use App\Domain\Repository\Exception\CategoryNotFoundException;
+use Ramsey\Uuid\Uuid;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Exception\BadRequestException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 #[Route('/blogs')]
 class BlogController extends AbstractController
 {
     public function __construct(
 		private BlogRepositoryInterface $blogRepository,
-        private SerializerInterface $serializer
+        private SerializerInterface $serializer,
+		private ValidatorInterface $validator
     ) {}
 
     #[Route('/', methods: 'GET')]
@@ -44,23 +48,23 @@ class BlogController extends AbstractController
             'json'
         );
 
-        try {
-            $resultBlogId = $blogCreateHandler->handle($blogCreateCommand);
-            return $this->json(['blogId' => $resultBlogId], Response::HTTP_CREATED);
-        } catch (CategoryNotFoundException $e) {
-            return $this->json(['error' => $e->getMessage()], Response::HTTP_NOT_FOUND);
-        }
+		$violations = $this->validator->validate($blogCreateCommand);
+		if ($violations->count() > 0) {
+			throw new BadRequestException();
+		}
+
+		$resultBlogId = $blogCreateHandler->handle($blogCreateCommand);
+		return $this->json(['blogId' => $resultBlogId], Response::HTTP_CREATED);
     }
 
     #[Route('/{id}', methods: 'GET')]
     public function getBlog(string $id): Response
     {
-        try {
-            $blog = $this->blogRepository->get(BlogId::generate($id));
-			return $this->json($blog, context: [AbstractNormalizer::GROUPS => ['rest']]);
-        } catch (BlogNotFoundException $e) {
-            return $this->json(['error' => $e->getMessage()], Response::HTTP_NOT_FOUND);
-        }
+		if (!Uuid::isValid($id)) {
+			throw new BadRequestException();
+		}
+		$blog = $this->blogRepository->get(BlogId::generate($id));
+		return $this->json($blog, context: [AbstractNormalizer::GROUPS => ['rest']]);
     }
 
     #[Route('/{id}/edit', methods: 'PUT')]
@@ -72,55 +76,40 @@ class BlogController extends AbstractController
             'json'
         );
 		$blogEditCommand->blogId = $id;
-        try {
-            $blogEditHandler->handle($blogEditCommand);
-            return $this->json([], Response::HTTP_ACCEPTED);
-        } catch (\DomainException $e) {
-            return $this->json(['error' => $e->getMessage()], Response::HTTP_NOT_FOUND);
-        }
+
+		$violations = $this->validator->validate($blogEditCommand);
+		if ($violations->count() > 0) {
+			throw new BadRequestException();
+		}
+
+		$blogEditHandler->handle($blogEditCommand);
+		return $this->json([], Response::HTTP_ACCEPTED);
     }
 
     #[Route('/{id}/delete', methods: 'DELETE')]
     public function removeBlog(string $id): Response
     {
-        try {
-			$blog = $this->blogRepository->get(BlogID::generate($id));
-			$this->blogRepository->remove($blog);
-            return $this->json([]);
-        } catch (BlogNotFoundException $e) {
-            return $this->json(['warning' => $e->getMessage()], Response::HTTP_NOT_FOUND);
-        }
+		if (!Uuid::isValid($id)) {
+			throw new BadRequestException();
+		}
+
+		$blog = $this->blogRepository->get(BlogID::generate($id));
+		$this->blogRepository->remove($blog);
+		return $this->json([]);
     }
 
 	#[Route('/{id}/articles', methods: ['GET'])]
 	public function getArticles(string $id): Response
 	{
-		try {
-			$blog = $this->blogRepository->get(BlogId::generate($id));
-			$articles = $blog->getArticles();
-			return $this->json($articles, context: [
-				AbstractNormalizer::GROUPS => ['rest'],
-				AbstractNormalizer::IGNORED_ATTRIBUTES => ['content']
-			]);
-		} catch (BlogNotFoundException $e) {
-			return $this->json(['error' => $e->getMessage()], Response::HTTP_NOT_FOUND);
+		if (!Uuid::isValid($id)) {
+			throw new BadRequestException();
 		}
-	}
 
-	#[Route('/{id}/publish', methods: ['POST'])]
-	public function publishArticle(string $id, Request $request, ArticlePublisherHandle $articlePublisherHandle): Response
-	{
-		try {
-			$articlePublishCommand = $this->serializer->deserialize(
-				$request->getContent(),
-				ArticlePublishCommand::class,
-				'json');
-
-			$articlePublishCommand->blogId = $id;
-			$article = $articlePublisherHandle->handle($articlePublishCommand);
-			return $this->json(['articleId' => $article->getId()], Response::HTTP_CREATED);
-		} catch (\DomainException $e) {
-			return $this->json(['error' => $e->getMessage()], Response::HTTP_NOT_FOUND);
-		}
+		$blog = $this->blogRepository->get(BlogId::generate($id));
+		$articles = $blog->getArticles();
+		return $this->json($articles, context: [
+			AbstractNormalizer::GROUPS => ['rest'],
+			AbstractNormalizer::IGNORED_ATTRIBUTES => ['content']
+		]);
 	}
 }
